feat: add monet generation in preset.py
