+++ gradience/commands/apply_preset.py
@@ -3,6 +3,7 @@ import click
 from gradience.modules.preset import Preset
 from gradience.modules.themes import Theme
 
+
 @click.command()
 @click.argument("preset", type=click.Path(exists=True), required=True)
 def apply_preset(preset):
