+++ po/POTFILES
@@ -22,6 +22,8 @@ gradience/modules/repo.py
 gradience/modules/run_async.py
 gradience/modules/shell.py
 gradience/modules/utils.py
+gradience/modules/exceptions.py
+gradience/modules/themes.py
 gradience/app_type_dialog.py
 gradience/builtin_preset_row.py
 gradience/custom_css_group.py
